// src/app/models/backend/game-list.model.ts
export class GameList {
    id?: any;
    game_image?: any;
    game_name?: any;
    game_description?: any;
    game_year?: any;
    game_platform?: any;
    game_rating?: any;
    game_url?: any;
    esrb_rating?: any;
    pegi_rating?: any;
    pegi_cd?: any;
    media?: any;
    URI?: any;
    image_path?: any;
    createdAt?: any;
}